test: common tests entities are in inuser_common_test.go
