+++ internal/web/controllers/user_add_test.go
@@ -3,13 +3,8 @@ package controllers_test
 import (
 	"bytes"
 	"encoding/json"
-	"errors"
-	"fmt"
 	"net/http"
-	"net/http/httptest"
-	"regexp"
 	"testing"
-	"time"
 
 	"github.com/golang/mock/gomock"
 	"github.com/goofinator/usersHttp/internal/repositories/mocks"
@@ -17,31 +12,7 @@ import (
 	"github.com/goofinator/usersHttp/internal/web/model"
 )
 
-var someError = errors.New("some error")
-
-var (
-	jsonValidStr = []byte(fmt.Sprintf(`{"Id": 0,
-	"Name": "petya",
-	"Lastname": "Pupkin",
-	"Age": 22,
-	"Birthdate": "%v"}`,
-		time.Now().UTC().Format(time.RFC3339)))
-	jsonInvalidStr = []byte(`{"Id": 0,
-	"Name": "petya",
-	"Lastname": "Pupkin",
-	"Age": 22,
-	"Birthdate": "AAA"}`)
-)
-
-type commonTestCase struct {
-	name       string
-	jsonStr    []byte
-	wantStatus int
-	wantBodyRE string
-	mockRet    error
-}
-
-var tests = []*commonTestCase{
+var testsAddUser = []*commonTestCase{
 	{
 		name:       "broken json",
 		jsonStr:    jsonInvalidStr,
@@ -53,36 +24,34 @@ var tests = []*commonTestCase{
 		name:       "db error",
 		jsonStr:    jsonValidStr,
 		wantStatus: http.StatusInternalServerError,
-		wantBodyRE: "",
+		wantBodyRE: "^error on AddUser: some error",
 		mockRet:    someError,
 	},
 	{
 		name:       "success",
 		jsonStr:    jsonValidStr,
 		wantStatus: http.StatusOK,
-		wantBodyRE: "",
+		wantBodyRE: "^$",
 		mockRet:    nil,
 	},
 }
 
 func TestAddUserHandler(t *testing.T) {
-	for _, test := range tests {
+	for _, test := range testsAddUser {
 		t.Run(test.name, func(t *testing.T) {
 			//You need a Storager mock to process the request
 			controller := gomock.NewController(t)
 			defer controller.Finish()
 			db := mocks.NewMockStorager(controller)
 
-			if bytes.Equal(test.jsonStr, jsonValidStr) {
-				setAddUserExpectations(db, test)
-			}
+			setAddUserExpectations(db, test)
 
 			req, err := http.NewRequest("POST", "/users", bytes.NewBuffer(test.jsonStr))
 			if err != nil {
 				t.Fatalf("unexpected fail of NewRequest: %s", err)
 			}
 			req.Header.Set("Content-Type", "application/json")
-			rr := handleRequest(req, db)
+			rr := handleRequest(req, db, AddUserHandler)
 
 			checkStatus(t, test.wantStatus, rr.Code)
 
@@ -92,34 +61,13 @@ func TestAddUserHandler(t *testing.T) {
 }
 
 func setAddUserExpectations(db *mocks.MockStorager, test *commonTestCase) {
+	if test.name == "broken json" {
+		return
+	}
+
 	var user model.User
 	json.Unmarshal(jsonValidStr, &user)
 
 	db.EXPECT().
 		AddUser(&user).Return(test.mockRet)
-}
-
-func handleRequest(req *http.Request, db *mocks.MockStorager) *httptest.ResponseRecorder {
-	rr := httptest.NewRecorder()
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		AddUserHandler(w, r, db)
-	})
-	handler.ServeHTTP(rr, req)
-	return rr
-}
-
-func checkStatus(t *testing.T, want, got int) {
-	if want != got {
-		t.Errorf("unexpected status code:\nwant: %v\ngot: %v", want, got)
-	}
-}
-
-func checkBodyByRE(t *testing.T, wantRE, got string) {
-	matched, err := regexp.MatchString(wantRE, got)
-	if err != nil {
-		t.Fatalf("unexpected fail on MatchString: %v", err)
-	}
-	if !matched {
-		t.Errorf("unexpected response body:\nwant: match to regexp: %q\ngot: %q", wantRE, got)
-	}
 }
